controller updates

+++ ds2024Team77/src/main/java/gr/hua/dit/ds/ds2024Team77/repository/MessagesRepository.java
@@ -5,5 +5,6 @@ import gr.hua.dit.ds.ds2024Team77.entities.Messages;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MessagesRepository extends JpaRepository<Messages, Integer> {
+public interface MessagesRepository extends JpaRepository<Messages, Long> {
+
 }
